🐛 FIX : add try-catch in chat user
